test: migrate tests to bun
