+++ package.json
@@ -18,9 +18,8 @@
     "postinstall": "bun build-icons.js",
     "lint": "prettier . --check",
     "lint:fix": "prettier . --write",
-    "test": "vitest run",
-    "test:watch": "vitest",
-    "test:cov": "vitest run --coverage",
+    "test": "bun test",
+    "test:cov": "bun test --coverage",
     "dev": "run-p -rl build:dev serve",
     "build": "vite build",
     "build:dev": "vite build --watch --mode development --minify false",
@@ -32,7 +31,6 @@
     "@types/leaflet": "^1.9.12",
     "@types/leaflet-fullscreen": "^1.0.9",
     "@types/webextension-polyfill": "^0.10.7",
-    "@vitest/coverage-v8": "^1.6.0",
     "npm-run-all": "^4.1.5",
     "oslllo-svg2": "^2.0.2",
     "prettier": "^3.3.2",
@@ -40,7 +38,6 @@
     "release-it-changelogen": "^0.1.0",
     "vite": "^5.3.2",
     "vite-plugin-static-copy": "^1.0.5",
-    "vitest": "^1.6.0",
     "web-ext": "^8.2.0",
     "webextension-polyfill": "^0.12.0"
   },
